fix: 글래너 제목이 너무 길어서 화면 깨지는 현상 수정, 수정 시 글자 제한

See merge request s06-webmobile1-sub2/S06P12A606!136

+++ frontend/glanner/src/Components/GroupPlannerMoreBtn.js
@@ -8,16 +8,28 @@ export const GroupPlannerMoreBtn = ({setHeadTitle}) => {
     const { pathname } = useLocation();
     const navigator = useNavigate();
     const deleteGlanner = () => {
-        const id = pathname.slice(7)
-        axios(`/api/glanner/${id}`, {method: 'DELETE'})
-            .then(res => {
-                alert(`${res.data.glannerName} 글래너가 삭제되었습니다.`)
-                navigator(`/`)
-            })
-            .catch(err => console.log(err))
+        const ok = window.confirm('정말 삭제하시겠습니까? 삭제된 데이터는 복구할 수 없습니다.')
+        if (ok) {
+            const id = pathname.slice(7)
+            axios(`/api/glanner/${id}`, {method: 'DELETE'})
+                .then(res => {
+                    alert(`글래너가 삭제되었습니다.`)
+                    navigator(`/`)
+                })
+                .catch(err => console.log(err))
+        }
     }
     const editGlanner = () => {
         const newName = window.prompt("변경하려는 글래너 이름을 쓰세요.")
+        if (newName.length > 20) {
+            alert('20자 이내로 작성해주세요.')
+            return
+        }
+
+        if (newName.trim() == '') {
+            alert('내용을 작성해주세요.')
+            return
+        }
         const id = pathname.slice(7)
         axios(`/api/glanner/`, {method: 'PUT', data: {glannerId: id, glannerName: newName}})
             .then(res => setHeadTitle(newName))
